Be more explicit about what we want to load
LoadByName without a Queue argument just looks for the first CF with the
requested Name. This could be a CF on another queue or global or on this
queue. Instead, look for the current queue and then look for global, but
don't find an identically named CF which is assigned to a completely
different Queue.