Add configuration hooks for form labels and input fields
FormTools creates tables for each form element it creates
and you sometimes want to add some text next to the label
that is not part of the label, like a link or footnote.
Likewise, you sometimes want to add some additional text after
the input field in the table created for the input.

This commit adds two configuration variables that can accept HTML to
add this additional text.