Reset context object when attempting to load CF
When attempting to load a CF, the code attempts to find a
global CF before loading by name alone. The LoadByName CF method
automatically sets the context object to global on that call,
so the subsequent call in Field doesn't find non-global CFs since
it is using the same CF object. Reset the context object to the
ticket so the name lookup can find non-global CFs.