defer to EmailInput default Size if cols is undefined
EmailInput already has a default Size value if it is not passed in -
currently at '40'. Deferring to that value allows for a more consistent
look to EmailInput elements created via FormTools without duplicating the
constant value (40).

Check exists $ARGS{cols} to allow for explicitly passing cols => undef,
which produces an input with no "size" set.