Clarify that we are clearing the context object
Make it clear that we are clearing the context object that
had been set to global context. We aren't setting the context
to the ticket because in the final case where LoadByName
is called, $ticket is undef.

This workaround is only needed for RT 4.2.0 - 4.2.4. LoadByName
is refactored in 4.2.5 to correctly set the context object.