Remove queue object from EditCustomField call
The Object passed to EditCustomField should be the type the
custom field is applied to. Passing the queue object would only
be correct for queue cfs, which is not a typical case. For ticket
cfs, the typical case, passing the queue object creates an incorrect
name in the form and the cf isn't processed properly on submit.

EditCustomField can generate the correct form element from the
passed CustomField object.

This issue has likely not been noticed because the queue object is
only set when you pass it via 'notes'.