Customize RTs Abort to prevent permission errors from CreateTicket
By default, CreateTicket calls Abort at the end if the current
user doesn't have ShowTicket. In FormTools, we aren't showing
the ticket at the end, we're showing the last configured form
page. Since end users are using a form wizard to create one
or more tickets, it's possible they don't have rights to directly
see the newly created ticket, especially in Self Service.

Add an exception to Abort so we can get the ticket object if
the create was successful, even if the current user can't
see the ticket.