Tweak the default value input
Previously it was a plain text input for all the fields, this commit
tweaks it to use textarea for core Content and HTML custom fields. For all
the other custom fields, it calls the standard EditCustomField to render
different inputs based on the custom field type.

With this change, now we can support multiple default values.

For combo/autocomplete custom fields, some enhanced js is not functioning
right after these fields are added(via dragging&dropping) right now.
Datepicker's position behaves weirdly in modal, so it's been disabled
intentionally. Note that this doesn't prevent users from setting default
values and it's not worse than previous plain text inputs anyway, so it's ok
to call EditCustomField consistently for these types.