Add to list of core fields supported by FomTools
Of these new fields, only Owner can be added to a form. Even then, it
doesn't allow for selection or autocompletion. By adding it to the list
of core fields, it also fixes a bug where Owner gets added as a core
field, but saved as a CustomField.

Next up is to add Due and TimeWorked to the form builder UI, and make
these fields render with the proper format.