Handle correctly setting 'required' when multiple checkboxes selected
When a customfield has a render type of 'List' we render a list of checkboxes,
when the extension's JS validates if a field is required or not it checks only the first
value found for this list. The JS should actually loop over the checked items in the case
that multiple boxes are selected.