Check for ContentType before ticket create
If the Content field is not on the last page of the form the ContentType arg
will not be set when the form is submitted.

Added a check to see if the ContentType arg needs to be set before creating the
ticket to ensure Content is formatted correctly.

See f0c081b42e for why ContentType is not included as a hidden form field.