Don't persist automatic form values as hidden form fields
The ContentType and create_ticket fields are added to form
pages automatically when the page is rendered, so they don't
need to be persisted as hidden values to save the values like
fields holding user inputs. Skip them when stashing other hidden
form data.

When they are added as hidden fields, subsequent submitted form
values present these as an arrayref containing the duplicate
values. This causes various errors for some processing that
expects a single value. For example, during Content processing,
text/html content is not properly escaped, causing raw HTML to
show up in the editor.

9bb5ca7698 previously fixed this by removing duplicates when
writing hidden values, which fixed the issue when using the Back
and Next buttons. But adding attachments causes the same page to
be submitted to itself, allowing the duplicate values to show up
in %ARGS in earlier processing.

Since these values don't need to be persisted, we shouldn't write
them to the page as "persisted" values at all.